django-backend/src/requirements.txt
bcrypt==3.1.7
boto3==1.9.227
botocore==1.12.227
Django==2.2.2
django-cors-headers==3.1.0
django-environ==0.4.5
django-filter==2.2.0
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
psycopg2==2.8.3
requests==2.22.0


